Use Java long type for integers sent to Stackdriver

+++ src/main/java/org/apache/druid/emitter/stackdriver/StackdriverMetricTimeseries.java
@@ -138,7 +138,7 @@ class StackdriverMetricTimeseriesSerializer extends StdSerializer<StackdriverMet
             jgen.writeEndObject();
 
             jgen.writeObjectFieldStart("value");
-            jgen.writeStringField("int64Value", String.valueOf(point.getValue().intValue()));
+            jgen.writeStringField("int64Value", String.valueOf(point.getValue().longValue()));
             jgen.writeEndObject();
 
             jgen.writeEndObject();
